% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistPlotter-package.R
\docType{package}
\name{DistPlotter-package}
\alias{DistPlotter-package}
\alias{_PACKAGE}
\alias{DistPlotter}
\title{DistPlotter: A Graphical User Interface for Plotting Common Univariate
  Distributions}
\description{
The DistPlotter package enables users to plot common univariate
  distributions and was designed with a particular focus on education. Users
  can plot distributions based on general shape (e.g., symmetric vs. skewed
  right) and distributions from common discrete and continuous families. They
  can also shade areas underneath the curve (e.g., areas corresponding to
  general probabilities, p-values, or confidence levels). Users can
  additionally plot their own quantitative data, as well as overlay a normal
  density curve and shade area(s) underneath.
}
\details{
Package: DistPlotter \cr
Type: Package \cr
Version: 0.0.2 \cr
Date: 2022-03-09 \cr
Depends: R (>= 3.5.0) \cr
Imports: colourpicker, dplyr, DT, extraDistr, ggplot2, rio, scales, shiny,
shinyalert, shinyalert, shinyBS, shinyjs, shinyWidgets, stringi, stringr
License: MIT \cr
BugReports: https://github.com/ccasement/DistPlotter/issues \cr
Encoding: UTF-8 \cr
}
\section{Function}{
 \itemize{
\item \code{\link{runDistPlotterApp}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ccasement/DistPlotter}
  \item \url{https://CRAN.R-project.org/package=DistPlotter}
  \item Report bugs at \url{https://github.com/ccasement/DistPlotter/issues}
}

}
\author{
Christopher Casement \cr
Department of Mathematics \cr
Fairfield University \cr
\email{casementc@gmail.com}

Grant Innerst \cr
Department of Mathematics \cr
Shippensburg University

Melissa Innerst \cr
Department of Mathematics \cr
Juniata College
}
