% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_plots.R
\name{plot_chisq}
\alias{plot_chisq}
\title{绘制卡方分布 (Chi-Squared)}
\usage{
plot_chisq(df = 3, fill_color = "orange")
}
\arguments{
\item{df}{数值型。自由度 (Degrees of Freedom)。}

\item{fill_color}{字符型。填充颜色。}
}
\value{
ggplot 对象
}
\description{
绘制卡方分布 (Chi-Squared)
}
\examples{
plot_chisq(df = 3)
plot_chisq(df = 10, fill_color = "purple")
}
