% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot_normal}
\alias{plot_normal}
\title{绘制正态分布密度图}
\usage{
plot_normal(mu = 0, sigma = 1, fill_color = "skyblue")
}
\arguments{
\item{mu}{数值型。正态分布的均值，默认为 0。}

\item{sigma}{数值型。正态分布的标准差，默认为 1。}

\item{fill_color}{字符型。分布区域的填充颜色，默认为 "skyblue"。}
}
\value{
返回一个 ggplot 对象。
}
\description{
这是一个基于 ggplot2 的函数，用于快速可视化正态分布。
}
\examples{
plot_normal(mu = 0, sigma = 1, fill_color = "orange")
plot_normal(mu = 10, sigma = 5)
}
