% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_plots.R
\name{plot_t_vs_normal}
\alias{plot_t_vs_normal}
\title{对比 T 分布与标准正态分布}
\usage{
plot_t_vs_normal(df = 1)
}
\arguments{
\item{df}{数值型。T 分布的自由度。自由度越小，尾部越厚。}
}
\value{
ggplot 对象
}
\description{
用来展示 T 分布的“厚尾”特性。红色为 T 分布，灰色虚线为标准正态分布。
}
\examples{
plot_t_vs_normal(df = 2) # 差异明显
plot_t_vs_normal(df = 30) # 几乎重合
}
