% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDistPlotterApp.R
\name{runDistPlotterApp}
\alias{runDistPlotterApp}
\title{Run the DistPlotter Shiny application}
\usage{
runDistPlotterApp()
}
\value{
There is no return value.
}
\description{
Runs the DistPlotter Shiny application.
}
\examples{
## only run the app in an interactive R session
if (interactive()) {runDistPlotterApp()}

}
\author{
Christopher Casement \cr
Department of Mathematics \cr
Fairfield University \cr
\email{casementc@gmail.com}

Grant Innerst \cr
Department of Mathematics \cr
Shippensburg University

Melissa Innerst \cr
Department of Mathematics \cr
Juniata College
}
